requirements.txt
numpy
tensorflow  
keras
opencv-python
matplotlib
tqdm
streamlit
pillow
regex